menu_lst: direct int-to-float for icon v offset, fuses fsubs (96.1%)

+++ src/menu_lst.cpp
@@ -88,7 +88,7 @@ void CMenuPcs::MLstDraw()
 			float iconY = (float)(item->y - 6);
 			v = zero;
 			if ((menuMode == 1) && (i == this->m_menuLstState->cursor)) {
-				v += (float)((double)item->height);
+				v += (float)item->height;
 			}
 			MenuPcs.DrawRect(
 				0,
